requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
hl7apy==1.3.4
fhir.resources==6.5.0